move metamask state selectors out of send
