+++ ui/hooks/useEthFiatAmount.js
@@ -1,12 +1,9 @@
 import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
-import {
-  getConversionRate,
-  getCurrentCurrency,
-  getShouldShowFiat,
-} from '../selectors';
+import { getCurrentCurrency, getShouldShowFiat } from '../selectors';
 import { decEthToConvertedCurrency } from '../helpers/utils/conversions.util';
 import { formatCurrency } from '../helpers/utils/confirm-tx.util';
+import { getConversionRate } from '../ducks/metamask/metamask';
 
 /**
  * Get an Eth amount converted to fiat and formatted for display
