+++ ui/hooks/useTransactionDisplayData.test.js
@@ -7,11 +7,10 @@ import transactions from '../../test/data/transaction-data.json';
 import {
   getPreferences,
   getShouldShowFiat,
-  getNativeCurrency,
   getCurrentCurrency,
   getCurrentChainId,
 } from '../selectors';
-import { getTokens } from '../ducks/metamask/metamask';
+import { getTokens, getNativeCurrency } from '../ducks/metamask/metamask';
 import { getMessage } from '../helpers/utils/i18n-helper';
 import messages from '../../app/_locales/en/messages.json';
 import { ASSET_ROUTE, DEFAULT_ROUTE } from '../helpers/constants/routes';
