+++ ui/pages/confirm-approve/confirm-approve.js
@@ -14,13 +14,12 @@ import {
   getTokenValueParam,
 } from '../../helpers/utils/token-util';
 import { useTokenTracker } from '../../hooks/useTokenTracker';
-import { getTokens } from '../../ducks/metamask/metamask';
+import { getTokens, getNativeCurrency } from '../../ducks/metamask/metamask';
 import {
   transactionFeeSelector,
   txDataSelector,
   getCurrentCurrency,
   getDomainMetadata,
-  getNativeCurrency,
   getUseNonceField,
   getCustomNonceValue,
   getNextSuggestedNonce,
@@ -28,6 +27,7 @@ import {
   getIsEthGasPriceFetched,
   getIsMainnet,
 } from '../../selectors';
+
 import { currentNetworkTxListSelector } from '../../selectors/transactions';
 import Loading from '../../components/ui/loading-screen';
 import { getCustomTxParamsData } from './confirm-approve.util';
