+++ ui/pages/send/send-content/send-amount-row/send-amount-row.container.js
@@ -1,6 +1,5 @@
 import { connect } from 'react-redux';
 import {
-  getConversionRate,
   getGasTotal,
   getPrimaryCurrency,
   getSendToken,
@@ -13,6 +12,7 @@ import {
 import { getAmountErrorObject, getGasFeeErrorObject } from '../../send.utils';
 import { setMaxModeTo, updateSendAmount } from '../../../../store/actions';
 import { updateSendErrors } from '../../../../ducks/send/send.duck';
+import { getConversionRate } from '../../../../ducks/metamask/metamask';
 import SendAmountRow from './send-amount-row.component';
 
 export default connect(mapStateToProps, mapDispatchToProps)(SendAmountRow);
