+++ ui/pages/swaps/swaps-gas-customization-modal/swaps-gas-customization-modal.container.js
@@ -8,9 +8,9 @@ import {
   getDefaultActiveButtonIndex,
   getRenderableGasButtonData,
   getUSDConversionRate,
-  getNativeCurrency,
   getSwapsDefaultToken,
 } from '../../../selectors';
+import { getNativeCurrency } from '../../../ducks/metamask/metamask';
 
 import {
   getSwapsCustomizationModalPrice,
