+++ ui/pages/swaps/view-quote/view-quote.js
@@ -38,12 +38,13 @@ import {
   getTokenExchangeRates,
   getSwapsDefaultToken,
   getCurrentChainId,
-  getNativeCurrency,
   isHardwareWallet,
   getHardwareWalletType,
 } from '../../../selectors';
+import { getNativeCurrency, getTokens } from '../../../ducks/metamask/metamask';
+
 import { toPrecisionWithoutTrailingZeros } from '../../../helpers/utils/util';
-import { getTokens } from '../../../ducks/metamask/metamask';
+
 import {
   safeRefetchQuotes,
   setCustomApproveTxData,
