+++ ui/selectors/send.test.js
@@ -1,9 +1,5 @@
 import sinon from 'sinon';
-import { TRANSACTION_STATUSES } from '../../shared/constants/transaction';
 import {
-  getBlockGasLimit,
-  getConversionRate,
-  getNativeCurrency,
   getGasLimit,
   getGasPrice,
   getGasTotal,
@@ -17,12 +13,9 @@ import {
   getSendFrom,
   getSendFromBalance,
   getSendFromObject,
-  getSendHexDataFeatureFlagState,
   getSendMaxModeState,
   getSendTo,
-  getSendToAccounts,
   getTokenBalance,
-  getUnapprovedTxs,
   gasFeeIsInError,
   getGasLoadingError,
   getGasButtonGroupShown,
@@ -84,18 +77,6 @@ describe('send selectors', () => {
     });
   });
 
-  describe('getBlockGasLimit', () => {
-    it('should return the current block gas limit', () => {
-      expect(getBlockGasLimit(mockState)).toStrictEqual('0x4c1878');
-    });
-  });
-
-  describe('getConversionRate()', () => {
-    it('should return the eth conversion rate', () => {
-      expect(getConversionRate(mockState)).toStrictEqual(1200.88200327);
-    });
-  });
-
   describe('getCurrentAccountWithSendEtherInfo()', () => {
     it('should return the currently selected account with identity info', () => {
       expect(getCurrentAccountWithSendEtherInfo(mockState)).toStrictEqual({
@@ -105,12 +86,6 @@ describe('send selectors', () => {
         address: '0xd85a4b6a394794842887b8284293d69163007bbb',
         name: 'Send Account 4',
       });
-    });
-  });
-
-  describe('getNativeCurrency()', () => {
-    it('should return the ticker symbol of the selected network', () => {
-      expect(getNativeCurrency(mockState)).toStrictEqual('ETH');
     });
   });
 
@@ -207,12 +182,6 @@ describe('send selectors', () => {
     });
   });
 
-  describe('getSendHexDataFeatureFlagState()', () => {
-    it('should return the sendHexData feature flag state', () => {
-      expect(getSendHexDataFeatureFlagState(mockState)).toStrictEqual(true);
-    });
-  });
-
   describe('getSendFrom()', () => {
     it('should return the send.from', () => {
       expect(getSendFrom(mockState)).toStrictEqual(
@@ -279,75 +248,9 @@ describe('send selectors', () => {
     });
   });
 
-  describe('getSendToAccounts()', () => {
-    it('should return an array including all the users accounts and the address book', () => {
-      expect(getSendToAccounts(mockState)).toStrictEqual([
-        {
-          code: '0x',
-          balance: '0x47c9d71831c76efe',
-          nonce: '0x1b',
-          address: '0xfdea65c8e26263f6d9a1b5de9555d2931a33b825',
-          name: 'Send Account 1',
-        },
-        {
-          code: '0x',
-          balance: '0x37452b1315889f80',
-          nonce: '0xa',
-          address: '0xc5b8dbac4c1d3f152cdeb400e2313f309c410acb',
-          name: 'Send Account 2',
-        },
-        {
-          code: '0x',
-          balance: '0x30c9d71831c76efe',
-          nonce: '0x1c',
-          address: '0x2f8d4a878cfa04a6e60d46362f5644deab66572d',
-          name: 'Send Account 3',
-        },
-        {
-          code: '0x',
-          balance: '0x0',
-          nonce: '0x0',
-          address: '0xd85a4b6a394794842887b8284293d69163007bbb',
-          name: 'Send Account 4',
-        },
-        {
-          address: '0x06195827297c7a80a443b6894d3bdb8824b43896',
-          name: 'Address Book Account 1',
-          chainId: '0x3',
-        },
-      ]);
-    });
-  });
-
   describe('getTokenBalance()', () => {
     it('should', () => {
       expect(getTokenBalance(mockState)).toStrictEqual(3434);
-    });
-  });
-
-  describe('getUnapprovedTxs()', () => {
-    it('should return the unapproved txs', () => {
-      expect(getUnapprovedTxs(mockState)).toStrictEqual({
-        4768706228115573: {
-          id: 4768706228115573,
-          time: 1487363153561,
-          status: TRANSACTION_STATUSES.UNAPPROVED,
-          gasMultiplier: 1,
-          metamaskNetworkId: '3',
-          txParams: {
-            from: '0xc5b8dbac4c1d3f152cdeb400e2313f309c410acb',
-            to: '0x18a3462427bcc9133bb46e88bcbe39cd7ef0e761',
-            value: '0xde0b6b3a7640000',
-            metamaskId: 4768706228115573,
-            metamaskNetworkId: '3',
-            gas: '0x5209',
-          },
-          txFee: '17e0186e60800',
-          txValue: 'de0b6b3a7640000',
-          maxCost: 'de234b52e4a0800',
-          gasPrice: '4a817c800',
-        },
-      });
     });
   });
